Test against k256 library
