+++ Cargo.toml
@@ -47,6 +47,7 @@ hex-literal = "1.0.0"
 either = "1.11.0"
 itertools = "0.14.0"
 jwt-simple = { version = "0.12.12", default-features = false, features = ["pure-rust"] }
+k256 = "0.13.4"
 lazy_static = "1.5.0"
 num-bigint = "0.4.6"
 num-integer = "0.1.46"
