+++ crates/frontend/Cargo.toml
@@ -22,6 +22,7 @@ smallvec.workspace = true
 base64 = { workspace = true }
 blake2 = { workspace = true }
 jwt-simple = { workspace = true }
+k256 = { workspace = true, features = ["arithmetic"] }
 rand = { workspace = true, features = ["std_rng", "thread_rng"] }
 rsa = { workspace = true, features = ["sha2"] }
 hex-literal = { workspace = true }
