+++ crates/frontend/src/circuits/ecdsa/mod.rs
@@ -2,7 +2,7 @@
 
 mod bitcoin;
 mod ecrecover;
-mod scalar_mul;
+pub mod scalar_mul;
 
 pub use bitcoin::verify as bitcoin_verify;
 pub use ecrecover::ecrecover;
